+++ clients/python/acn_client/__init__.py
@@ -44,7 +44,7 @@ from .models import (
 )
 from .realtime import ACNRealtime, ACNRealtimeOptions, AuthMode, WSState
 
-__version__ = "0.7.0"
+__version__ = "0.7.1"
 __all__ = [
     # Client
     "ACNClient",